indoor_drone_nav_v2/requirements.txt
# Core
numpy
pyyaml
pandas

# FastAPI Web Server
fastapi
uvicorn[standard]
websockets
python-socketio

# Testing
pytest
pytest-asyncio

# Deployment
docker
